requirements.txt
asgiref==3.11.1
awsebcli==3.22.1
botocore==1.35.99
cement==2.10.14
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
idna==3.11
jmespath==1.1.0
packaging==24.2
pathspec==0.10.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywin32==311
PyYAML==6.0.3
requests==2.32.5
semantic-version==2.10.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
termcolor==2.5.0
tzdata==2025.3
urllib3==1.26.20
wcwidth==0.2.14
